development ongoing

+++ LSGS/Views/FriendsListPage.xaml.cs
@@ -17,12 +17,18 @@ namespace LSGS.Views
         public ObservableCollection<Profile> friendsList { get; set; }
         Profile selectedProfile = new Profile();
         public Command DeclineCommand { get; }
+        public static bool IsRecommended = false;
         public FriendsListPage()
         {
             InitializeComponent();
             friendsList = new ObservableCollection<Profile>();
             CreateFriendsList();
             DeclineCommand = new Command(RemoveClicked);
+            if(IsRecommended)
+            {
+                RemoveButton.Text = "Recommend";
+                RemoveButton.BackgroundColor = Color.CadetBlue;
+            }
             BindingContext = this;
         }
 
@@ -39,16 +45,39 @@ namespace LSGS.Views
                 Globals.connection.Open();
             // create a DB command and set the SQL statement with parameters
             var command = Globals.connection.CreateCommand();
-            command.CommandText = $@"DELETE FROM Friends where (isAccepted=1 and User = {Globals.profile.METU_ID} and Friend = {selectedProfile.METU_ID});";
-            try
+         
+            if(IsRecommended)
             {
-                var insert = await command.ExecuteNonQueryAsync();
+                command.CommandText = $@"INSERT INTO Recommend (`Recommender`, `Recommendee`, `Book_ID`) 
+                  VALUES('{Globals.profile.METU_ID}', '{selectedProfile.METU_ID}', '{BookSearchResultsPage.BookSerialNo}');";
+                try
+                {
+                    var insert = await command.ExecuteNonQueryAsync();
+                    App.Current.MainPage.DisplayAlert("Success", $"Book is recommended to {selectedProfile.Name}!", "OK");
+                }
+                catch(Exception ex) 
+                { 
+                    if(ex.Message.Contains("Duplicate"))
+                        App.Current.MainPage.DisplayAlert("Error", $"Book is already recommended to {selectedProfile.Name}!", "OK");
+
+                }
+                finally { Globals.connection.Close(); }
             }
-            catch { }
-            finally { Globals.connection.Close(); }          
-            
-            App.Current.MainPage.DisplayAlert("Success", "Friend is removed!", "OK");
-            friendsList.Remove(selectedProfile);
+            else
+            {
+                command.CommandText = $@"DELETE FROM Friends where (isAccepted=1 and User = {Globals.profile.METU_ID} and Friend = {selectedProfile.METU_ID});";
+                try
+                {
+                    var insert = await command.ExecuteNonQueryAsync();
+                    App.Current.MainPage.DisplayAlert("Success", "Friend is removed!", "OK");
+                    friendsList.Remove(selectedProfile);
+                }
+                catch(Exception ex) 
+                {
+                    App.Current.MainPage.DisplayAlert("Error", "Friend cannot be removed.", "OK");
+                }
+                finally { Globals.connection.Close(); }
+            }
             //await Shell.Current.GoToAsync("ProfilePage");
         }
 
@@ -59,7 +88,7 @@ namespace LSGS.Views
                 Globals.connection.Open();
             // create a DB command and set the SQL statement with parameters
             var command = Globals.connection.CreateCommand();
-            command.CommandText = $"select * from User where METU_ID in (select Friend from Friends where isAccepted = 1 and(User = { Globals.profile.METU_ID})); ";
+            command.CommandText = $"SELECT * FROM User Where METU_ID in (SELECT Friend From Friends Where (User = {Globals.profile.METU_ID} and isAccepted=1)) UNION SELECT * FROM User Where METU_ID in (SELECT User From Friends Where(Friend = {Globals.profile.METU_ID} and isAccepted = 1)); ";
             try
             {
                 var reader = await command.ExecuteReaderAsync();
@@ -74,9 +103,10 @@ namespace LSGS.Views
                     friendsList.Add(new Profile(Name, Surname, METU_ID, Password, Email, Description));
                 }
             }
-            catch
+            catch(Exception ex)
             {
-
+                App.Current.MainPage.DisplayAlert("Error", "Friends list cannot be shown right now!", "OK");
+                await Navigation.PushAsync(new ProfilePage());
             }
             finally
             {
